# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GetVisitorDto": {
				"properties": {
					"visitorId": {
						"type": "number",
						"format": "double"
					},
					"visitorGuid": {
						"type": "string"
					},
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"visitorId",
					"visitorGuid",
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetVisitorDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetVisitorDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetVisitorDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GetVisitorDetailsDto": {
				"properties": {
					"visitorId": {
						"type": "number",
						"format": "double"
					},
					"visitorGuid": {
						"type": "string"
					},
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"visitorId",
					"visitorGuid",
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime",
					"status",
					"createdBy",
					"updatedBy",
					"createdDateTime",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorDetailsDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetVisitorDetailsDto"
							},
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorDetailsDto"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"IResponse_any_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{},
							{
								"items": {},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateVisitorDto": {
				"properties": {
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					}
				},
				"required": [
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EditVisitorByIdDto": {
				"properties": {
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					}
				},
				"required": [
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenderEnum": {
				"enum": [
					"M",
					"F"
				],
				"type": "string"
			},
			"RoleEnum": {
				"enum": [
					"SA",
					"STF",
					"RES",
					"SUB"
				],
				"type": "string"
			},
			"GetUserByAdminDto": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"userGuid": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"contactNumber": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/GenderEnum"
					},
					"role": {
						"$ref": "#/components/schemas/RoleEnum"
					},
					"userStatus": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userGuid",
					"firstName",
					"lastName",
					"userName",
					"contactNumber",
					"gender",
					"role",
					"userStatus",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetUserByAdminDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetUserByAdminDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetUserByAdminDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetUserByAdminDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetUserByAdminDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GeneralFileResponseDto": {
				"properties": {
					"fileGuid": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fileGuid",
					"fileName",
					"fileUrl",
					"contentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResidentInformationDto": {
				"properties": {
					"floor": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					},
					"supportedDocuments": {
						"items": {
							"$ref": "#/components/schemas/GeneralFileResponseDto"
						},
						"type": "array"
					}
				},
				"required": [
					"floor",
					"unit",
					"supportedDocuments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SystemInformationDto": {
				"properties": {
					"staffId": {
						"type": "string"
					},
					"supportedDocuments": {
						"items": {
							"$ref": "#/components/schemas/GeneralFileResponseDto"
						},
						"type": "array"
					}
				},
				"required": [
					"staffId",
					"supportedDocuments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUserDetailsByIdDto": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"userGuid": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"contactNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleEnum"
					},
					"roleInformation": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ResidentInformationDto"
							},
							{
								"$ref": "#/components/schemas/SystemInformationDto"
							}
						]
					},
					"dateOfBirth": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userGuid",
					"firstName",
					"lastName",
					"userName",
					"email",
					"contactNumber",
					"gender",
					"role",
					"dateOfBirth",
					"status",
					"createdBy",
					"createdDateTime",
					"updatedBy",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetUserDetailsByIdDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetUserDetailsByIdDto"
							},
							{
								"items": {
									"$ref": "#/components/schemas/GetUserDetailsByIdDto"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EditUserDetailsByIdDto": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"contactNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"userName",
					"email",
					"contactNumber",
					"gender",
					"dateOfBirth"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Secure Guard Pro",
		"version": "1.0.0",
		"description": "Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/visitors/admin/{id}/details": {
			"get": {
				"operationId": "getVisitorDetails",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor Management"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/visitors/create": {
			"post": {
				"operationId": "createVisitor",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVisitorDto"
							}
						}
					}
				}
			}
		},
		"/visitors": {
			"get": {
				"operationId": "getVisitorByResident",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetVisitorDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "isPast",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/visitors/details": {
			"get": {
				"operationId": "getVisitorDetails",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "visitorGuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/visitors/edit": {
			"put": {
				"operationId": "editVisitorById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "visitorGuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditVisitorByIdDto"
							}
						}
					}
				}
			}
		},
		"/users/admin": {
			"get": {
				"operationId": "getUserListByAdmin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetUserByAdminDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetUserByAdminDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "isActive",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": true,
						"schema": {}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/admin/{id}/details": {
			"get": {
				"operationId": "getUserDetailsById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetUserDetailsByIdDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/admin/{id}/activate": {
			"put": {
				"operationId": "activateUserById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/admin/{id}/deactivate": {
			"put": {
				"operationId": "deactivateUserById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/admin/{id}": {
			"delete": {
				"operationId": "deleteUserById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "createUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"newUser": [
							"RES",
							"SA",
							"SUB",
							"STF"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"put": {
				"operationId": "editUserProfileById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"SA",
							"RES",
							"SUB",
							"STF"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditUserDetailsByIdDto"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}